fix bugs, fix male, add mappers

+++ user-service/src/main/java/com/klasevich/itrex/lab/repository/impl/HibernateUserRepositoryImpl.java
@@ -1,16 +1,18 @@
 package com.klasevich.itrex.lab.repository.impl;
 
 import com.klasevich.itrex.lab.entity.User;
+import com.klasevich.itrex.lab.exception.RepositoryException;
 import com.klasevich.itrex.lab.repository.UserRepository;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.hibernate.Session;
+import org.hibernate.Transaction;
+import org.springframework.stereotype.Repository;
 
-import java.sql.SQLException;
 import java.util.List;
 
+@Repository
 public class HibernateUserRepositoryImpl implements UserRepository {
-    private static final Logger logger = LogManager.getLogger();
     private final Session session;
 
     public HibernateUserRepositoryImpl(Session session) {
@@ -28,17 +30,42 @@ public class HibernateUserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public void addAll(List<User> users) throws SQLException {
-        for (User user : users) {
-            session.save(user);
+    public void addAll(List<User> users) throws RepositoryException {
+        Transaction transaction = null;
+        try {
+            transaction = session.beginTransaction();
+            for (User user : users) {
+                session.save(user);
+            }
+            transaction.commit();
+        } catch (Exception e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
+            throw new RepositoryException("Transaction unable", e);
         }
     }
 
     @Override
-    public void delete(int id) {
-        User user = session.get(User.class, id);
-        logger.debug("user {} will be deleted with id {}", user, id);
-        session.delete(user);
+    public int delete(int id) throws RepositoryException {
+        Transaction transaction = null;
+        int result = id;
+        try {
+            transaction = session.beginTransaction();
+            User user = session.find(User.class, id);
+            if (user != null) {
+                session.delete(user);
+            } else {
+                result = 0;
+            }
+            transaction.commit();
+        } catch (Exception e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
+            throw new RepositoryException("Transaction unable", e);
+        }
+        return result;
     }
 
     @Override
